Flatten postgres package
When an application is small, subdividing repository packages
prematurely may make future development harder. For example, if you
discover that the data objects of one package are needed by another.
