+++ internal/adapter/repository/postgres/doc.go
@@ -1,0 +1,5 @@
+// Package postgres exposes operations on a Postgres-backed transfer
+// repository. It includes the following tables:
+//   * bank_accounts
+//   * transactions
+package postgres
