--- internal/adapter/repository/postgres/transferrepo/doc.go
@@ -1,5 +1,0 @@
-// Package transferrepo exposes operations on a Postgres-backed transfer
-// repository. It includes the following tables:
-//   * bank_accounts
-//   * transactions
-package transferrepo
