+++ internal/integration_test/integration_test.go
@@ -15,7 +15,6 @@ import (
 
 	"github.com/angusgmorrison/hexagonal/internal/adapter/envconfig"
 	"github.com/angusgmorrison/hexagonal/internal/adapter/repository/postgres"
-	"github.com/angusgmorrison/hexagonal/internal/adapter/repository/postgres/transferrepo"
 	"github.com/angusgmorrison/hexagonal/internal/adapter/rest"
 	server "github.com/angusgmorrison/hexagonal/internal/adapter/rest"
 	"github.com/angusgmorrison/hexagonal/internal/controller"
@@ -53,12 +52,12 @@ func NewServer() (*server.Server, error) {
 
 	envConfig := defaultEnvConfig()
 
-	db, err := postgres.New(envConfig.DB)
+	db, err := postgres.NewDB(envConfig.DB)
 	if err != nil {
 		return nil, fmt.Errorf("create database: %w", err)
 	}
 
-	transferRepo, err := transferrepo.New(db, envConfig.App)
+	transferRepo, err := postgres.NewTransferRepository(db, envConfig.App)
 	if err != nil {
 		return nil, fmt.Errorf("create transfer Repository: %w", err)
 	}
@@ -97,12 +96,12 @@ const (
 )
 
 type repository struct {
-	db *postgres.Postgres
+	db *postgres.DB
 }
 
 func (r *repository) insertBankAccount(
-	br transferrepo.BankAccountRow,
-) (transferrepo.BankAccountRow, error) {
+	br postgres.BankAccountRow,
+) (postgres.BankAccountRow, error) {
 	query, args, err := r.db.BindNamed(_insertBankAccountQuery, br)
 	if err != nil {
 		return br, fmt.Errorf("db.BindNamed: %w", err)
@@ -115,8 +114,8 @@ func (r *repository) insertBankAccount(
 	return br, nil
 }
 
-func (r *repository) getBankAccountByID(id int64) (transferrepo.BankAccountRow, error) {
-	var row transferrepo.BankAccountRow
+func (r *repository) getBankAccountByID(id int64) (postgres.BankAccountRow, error) {
+	var row postgres.BankAccountRow
 
 	err := r.db.Get(&row, _getBankAccountByIDQuery, id)
 	if err != nil {
@@ -156,8 +155,8 @@ func (r *repository) countTransactions() (int64, error) {
 
 func (r *repository) selectTransactionsByCounterpartyName(
 	name string,
-) (transferrepo.TransactionRows, error) {
-	var rows transferrepo.TransactionRows
+) (postgres.TransactionRows, error) {
+	var rows postgres.TransactionRows
 
 	if err := r.db.Select(&rows, _selectTransactionByCounterpartyNameQuery, name); err != nil {
 		return nil, fmt.Errorf("selectTransactionByCounterpartyName: %w", err)
@@ -180,7 +179,7 @@ type infrastructure struct {
 }
 
 func newInfrastructure(env envconfig.EnvConfig) (*infrastructure, error) {
-	db, err := postgres.New(env.DB)
+	db, err := postgres.NewDB(env.DB)
 	if err != nil {
 		return nil, fmt.Errorf("create database: %w", err)
 	}
@@ -248,8 +247,8 @@ func serverURL() string {
 	return fmt.Sprintf("http://0.0.0.0:%d", _serverPort)
 }
 
-func defaultBankAccount() transferrepo.BankAccountRow {
-	return transferrepo.BankAccountRow{
+func defaultBankAccount() postgres.BankAccountRow {
+	return postgres.BankAccountRow{
 		OrganizationName: "ACME Corp",
 		BIC:              "OIVUSCLQXXX",
 		IBAN:             "FR10474608000002006107XXXXX",
